backend/requirements.txt
# 肿瘤治疗进展查询系统 - 后端依赖
# Python 3.11+

# Web框架
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# 缓存和消息队列
redis==5.0.1
celery==5.3.4

# 搜索引擎
elasticsearch==8.11.1

# 数据处理和科学计算
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# 自然语言处理
spacy==3.7.2
nltk==3.8.1

# 机器学习（可选）
scikit-learn==1.3.2

# 爬虫相关
scrapy==2.11.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.4
selenium==4.16.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# 工具库
python-dateutil==2.8.2
arrow==1.3.0
uuid==1.30

# 日志和监控
loguru==0.7.2
prometheus-fastapi-instrumentator==6.1.0

# 测试
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.26.0

# 开发工具
black==23.12.1
flake8==6.1.0
mypy==1.8.0
isort==5.13.2

# 文档生成
mkdocs==1.5.3
mkdocs-material==9.5.2